+++ src/main/kotlin/org/berrycrush/intellij/parser/impl/ParameterParsers.kt
@@ -101,12 +101,11 @@ private fun PsiBuilder.tryParseParameterLike(
         BerryCrushTokenTypes.TEXT,
         -> if (parseParameterKey(marker)) return false
         else -> {
-            if (!allowBody) {
+            if (tokenType == BerryCrushTokenTypes.BODY && !allowBody) {
                 marker.rollbackTo()
                 return false
             }
 
-            val tokenType = tokenType
             if (tokenType == null || tokenType == BerryCrushTokenTypes.NEWLINE) {
                 marker.rollbackTo()
                 return false