+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - `parameters` block formatting
-- `if-else` directive formatting 
+- `if-else` directive formatting
+- PSI parsing for `parameters` entries.
 
 ## [2.0.1]